Revert original namespaced dashboard code and re-add validateDashboardLoad as non-breaking addition. (#15)
* Revert "Add support for validating dashboard load configuration and split models into separate report and dashboard modules. (#13)"

This reverts commit 64baf4ce46830c9b9e1c6fbfa8ef3e110e7ff09e.

* Add dashboard load configuration schema and validateDashboardLoad method as non-breaking addition.

* Change normalizeError method to be non-destructive.

* Change ILoadConfiguration to IReportLoadConfiguration and validateLoad to validateReportLoad to reduce confusion of being generic load configuration compared to dashboard load configuration.
