+++ src/schemas/dashboardLoadConfiguration.json
