+++ test/models.spec.ts
@@ -1,121 +1,23 @@
 import * as models from '../src/models';
 
-describe('Unit | models.dashboard', function () {
-
-    function testForExpectedMessage(errors: models.IError[], message: string) {
-      expect(errors).toBeDefined();
-      let flag = false;
-      errors.forEach(error => {
-        if (error.message === message) {
-            flag = true;
-        }
-      });
-	  
-	  expect(flag).toBe(true);
-    }
-  
-  describe('validateLoad', function () {
-
-    const accessTokenRequiredMessage = models.dashboard.loadSchema.properties.accessToken.messages.required;
-    const accessTokenInvalidTypeMessage = models.dashboard.loadSchema.properties.accessToken.invalidMessage;
-    const idRequiredMessage = models.dashboard.loadSchema.properties.id.messages.required;
-    const idInvalidTypeMessage = models.dashboard.loadSchema.properties.id.invalidMessage;
-
-    it(`should return errors with one containing message '${accessTokenRequiredMessage}' if accessToken is not defined`, function () {
-        // Arrange
-        const testData = {
-            load: {
-            }
-        };
-
-        // Act
-        const errors = models.dashboard.validateLoad(testData.load);
-
-        // Assert
-        testForExpectedMessage(errors, accessTokenRequiredMessage);
-    });
-
-    it(`should return errors with one containing message '${accessTokenInvalidTypeMessage}' if accessToken is not a string`, function () {
-        // Arrange
-        const testData = {
-            load: {
-                accessToken: 1
-            }
-        };
-
-        // Act
-        const errors = models.dashboard.validateLoad(testData.load);
-
-        // Assert
-        testForExpectedMessage(errors, accessTokenInvalidTypeMessage);
-    });
-
-    it(`should return errors with one containing message '${idRequiredMessage}' if id is not defined`, function () {
-        // Arrange
-        const testData = {
-            load: {
-                accessToken: '123'
-            }
-        };
-
-        // Act
-        const errors = models.dashboard.validateLoad(testData.load);
-
-        // Assert
-        testForExpectedMessage(errors, idRequiredMessage);
-    });
-
-    it(`should return errors with one containing message '${idInvalidTypeMessage}' if id is not a string`, function () {
-        // Arrange
-        const testData = {
-           load: {
-               id: 1
-           }
-        };
-
-        // Act
-        const errors = models.dashboard.validateLoad(testData.load);
-
-        // Assert
-        testForExpectedMessage(errors, idInvalidTypeMessage);
-    });
-
-    it(`should return undefined if id and accessToken are provided`, function () {
-        // Arrange
-        const testData = {
-            load: {
-               id: 'fakeId',
-               accessToken: 'fakeAccessToken'
-            }
-        };
-
-        // Act
-        const errors = models.dashboard.validateLoad(testData.load);
-
-        // Assert
-        expect(errors).toBeUndefined();
-     });
-  });
-});
-
-describe('Unit | models.report', function () {
+describe('Unit | Models', function () {
   function testForExpectedMessage(errors: models.IError[], message: string) {
     expect(errors).toBeDefined();
     errors
       .forEach(error => {
         if (error.message === message) {
-            expect(true).toBe(true);
+          expect(true).toBe(true);
         }
       });
   }
-  
-  describe('validateLoad', function () {
-    const accessTokenRequiredMessage = models.report.loadSchema.properties.accessToken.messages.required;
-    const accessTokenInvalidTypeMessage = models.report.loadSchema.properties.accessToken.messages.type;
-    const idRequiredMessage = models.report.loadSchema.properties.id.messages.required;
-    const idInvalidTypeMessage = models.report.loadSchema.properties.id.messages.type;
-    const filtersInvalidMessage = models.report.loadSchema.properties.filters.invalidMessage;
-    const pageNameInvalidTypeMessage = models.report.loadSchema.properties.pageName.messages.type;
+
+  describe('validateReportLoad', function () {
+    const accessTokenRequiredMessage = models.loadSchema.properties.accessToken.messages.required;
+    const accessTokenInvalidTypeMessage = models.loadSchema.properties.accessToken.messages.type;
+    const idRequiredMessage = models.loadSchema.properties.id.messages.required;
+    const idInvalidTypeMessage = models.loadSchema.properties.id.messages.type;
+    const filtersInvalidMessage = models.loadSchema.properties.filters.invalidMessage;
+    const pageNameInvalidTypeMessage = models.loadSchema.properties.pageName.messages.type;
 
     it(`should return errors with one containing message '${accessTokenRequiredMessage}' if accessToken is not defined`, function () {
       // Arrange
@@ -125,7 +27,7 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateLoad(testData.load);
+      const errors = models.validateReportLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, accessTokenRequiredMessage);
@@ -140,7 +42,7 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateLoad(testData.load);
+      const errors = models.validateReportLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, accessTokenInvalidTypeMessage);
@@ -154,7 +56,7 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateLoad(testData.load);
+      const errors = models.validateReportLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, idRequiredMessage);
@@ -168,7 +70,7 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateLoad(testData.load);
+      const errors = models.validateReportLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, idRequiredMessage);
@@ -184,7 +86,7 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateLoad(testData.load);
+      const errors = models.validateReportLoad(testData.load);
 
       // Assert
       expect(errors).toBeUndefined();
@@ -201,7 +103,7 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateLoad(testData.load);
+      const errors = models.validateReportLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, filtersInvalidMessage);
@@ -218,10 +120,90 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateLoad(testData.load);
+      const errors = models.validateReportLoad(testData.load);
 
       // Assert
       testForExpectedMessage(errors, pageNameInvalidTypeMessage);
+    });
+  });
+
+  describe('validateDashboardLoad', function () {
+    const accessTokenRequiredMessage = models.dashboardLoadSchema.properties.accessToken.messages.required;
+    const accessTokenInvalidTypeMessage = models.dashboardLoadSchema.properties.accessToken.messages.type;
+    const idRequiredMessage = models.dashboardLoadSchema.properties.id.messages.required;
+    const idInvalidTypeMessage = models.dashboardLoadSchema.properties.id.messages.type;
+
+    it(`should return errors with one containing message '${accessTokenRequiredMessage}' if accessToken is not defined`, function () {
+      // Arrange
+      const testData = {
+        load: {
+        }
+      };
+
+      // Act
+      const errors = models.validateDashboardLoad(testData.load);
+
+      // Assert
+      testForExpectedMessage(errors, accessTokenRequiredMessage);
+    });
+
+    it(`should return errors with one containing message '${accessTokenInvalidTypeMessage}' if accessToken is not a string`, function () {
+      // Arrange
+      const testData = {
+        load: {
+          accessToken: 1
+        }
+      };
+
+      // Act
+      const errors = models.validateDashboardLoad(testData.load);
+
+      // Assert
+      testForExpectedMessage(errors, accessTokenInvalidTypeMessage);
+    });
+
+    it(`should return errors with one containing message '${idRequiredMessage}' if id is not defined`, function () {
+      // Arrange
+      const testData = {
+        load: {
+        }
+      };
+
+      // Act
+      const errors = models.validateDashboardLoad(testData.load);
+
+      // Assert
+      testForExpectedMessage(errors, idRequiredMessage);
+    });
+
+    it(`should return errors with one containing message '${idInvalidTypeMessage}' if id is not a string`, function () {
+      // Arrange
+      const testData = {
+        load: {
+        }
+      };
+
+      // Act
+      const errors = models.validateDashboardLoad(testData.load);
+
+      // Assert
+      testForExpectedMessage(errors, idRequiredMessage);
+    });
+
+    it(`should return undefined if id and accessToken are provided`, function () {
+      // Arrange
+      const testData = {
+        load: {
+          id: 'fakeId',
+          accessToken: 'fakeAccessToken'
+        }
+      };
+
+      // Act
+      const errors = models.validateDashboardLoad(testData.load);
+
+      // Assert
+      expect(errors).toBeUndefined();
     });
   });
 
@@ -255,9 +237,9 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateFilter(malformedFilter);
-      const errors1 = models.report.validateFilter(malformedFilter1);
-      const errors2 = models.report.validateFilter(malformedFilter2);
+      const errors = models.validateFilter(malformedFilter);
+      const errors1 = models.validateFilter(malformedFilter1);
+      const errors2 = models.validateFilter(malformedFilter2);
 
       // Assert
       expect(errors).toBeDefined();
@@ -267,7 +249,7 @@ describe('Unit | models.report', function () {
 
     it("should return undefined if object is valid basic filter schema", function () {
       // Arrange
-      const expectedFilter: models.report.IBasicFilter = {
+      const expectedFilter: models.IBasicFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -282,18 +264,18 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const filter = new models.report.BasicFilter(
+      const filter = new models.BasicFilter(
         expectedFilter.target,
         expectedFilter.operator,
         expectedFilter.values);
 
       // Assert
-      expect(models.report.validateFilter(filter.toJSON())).toBeUndefined();
+      expect(models.validateFilter(filter.toJSON())).toBeUndefined();
     });
 
     it("should return undefined if object is valid advanced filter schema", function () {
       // Arrange
-      const expectedFilter: models.report.IAdvancedFilter = {
+      const expectedFilter: models.IAdvancedFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -316,19 +298,19 @@ describe('Unit | models.report', function () {
         ]
       };
 
-      const filter = new models.report.AdvancedFilter(
+      const filter = new models.AdvancedFilter(
         expectedFilter.target,
         expectedFilter.logicalOperator,
         ...expectedFilter.conditions.slice(0, 2));
 
-      const filter2 = new models.report.AdvancedFilter(
+      const filter2 = new models.AdvancedFilter(
         expectedFilter.target,
         expectedFilter.logicalOperator,
         ...expectedFilter.conditions.slice(1, 3));
 
       // Act
-      const errors = models.report.validateFilter(filter.toJSON());
-      const errors2 = models.report.validateFilter(filter2.toJSON());
+      const errors = models.validateFilter(filter.toJSON());
+      const errors2 = models.validateFilter(filter2.toJSON());
 
       // Assert
       expect(errors).toBeUndefined();
@@ -337,8 +319,8 @@ describe('Unit | models.report', function () {
   });
 
   describe('validateSettings', function () {
-    const filterPaneEnabledInvalidTypeMessage = models.report.settingsSchema.properties.filterPaneEnabled.messages.type;
-    const navContentPaneEnabledInvalidTypeMessage = models.report.settingsSchema.properties.navContentPaneEnabled.messages.type;
+    const filterPaneEnabledInvalidTypeMessage = models.settingsSchema.properties.filterPaneEnabled.messages.type;
+    const navContentPaneEnabledInvalidTypeMessage = models.settingsSchema.properties.navContentPaneEnabled.messages.type;
 
     it(`should return errors with one containing message '${filterPaneEnabledInvalidTypeMessage}' if filterPaneEnabled is not a boolean`, function () {
       // Arrange
@@ -349,7 +331,7 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateSettings(testData.settings);
+      const errors = models.validateSettings(testData.settings);
 
       // Assert
       testForExpectedMessage(errors, filterPaneEnabledInvalidTypeMessage);
@@ -364,7 +346,7 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateSettings(testData.settings);
+      const errors = models.validateSettings(testData.settings);
 
       // Assert
       testForExpectedMessage(errors, navContentPaneEnabledInvalidTypeMessage);
@@ -378,7 +360,7 @@ describe('Unit | models.report', function () {
       };
 
       // Act
-      const errors = models.report.validateSettings(testData.settings);
+      const errors = models.validateSettings(testData.settings);
 
       // Assert
       expect(errors).toBeUndefined();
@@ -392,7 +374,7 @@ describe("Unit | Filters", function () {
       // Arrange
 
       // Act
-      const basicFilter = new models.report.BasicFilter({ table: "t", column: "c" }, "In", 1, 2);
+      const basicFilter = new models.BasicFilter({ table: "t", column: "c" }, "In", 1, 2);
 
       // Assert
       expect(basicFilter.values).toEqual([1, 2]);
@@ -403,7 +385,7 @@ describe("Unit | Filters", function () {
       const values = [1, 2];
 
       // Act
-      const basicFilter = new models.report.BasicFilter({ table: "t", column: "c" }, "In", values);
+      const basicFilter = new models.BasicFilter({ table: "t", column: "c" }, "In", values);
 
       // Assert
       expect(basicFilter.values).toEqual(values);
@@ -411,7 +393,7 @@ describe("Unit | Filters", function () {
 
     it("should return valid json format when toJSON is called", function () {
       // Arrange
-      const expectedFilter: models.report.IBasicFilter = {
+      const expectedFilter: models.IBasicFilter = {
         $schema: "http://powerbi.com/product/schema#basic",
         target: {
           table: "a",
@@ -426,7 +408,7 @@ describe("Unit | Filters", function () {
       };
 
       // Act
-      const filter = new models.report.BasicFilter(
+      const filter = new models.BasicFilter(
         expectedFilter.target,
         expectedFilter.operator,
         expectedFilter.values);
@@ -437,7 +419,7 @@ describe("Unit | Filters", function () {
 
     it("can be constructed using either array form or individual arguments", function () {
       // Arrange
-      const expectedFilter: models.report.IBasicFilter = {
+      const expectedFilter: models.IBasicFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -452,8 +434,8 @@ describe("Unit | Filters", function () {
       };
 
       // Act
-      const filter1 = new models.report.BasicFilter(expectedFilter.target, expectedFilter.operator, expectedFilter.values);
-      const filter2 = new models.report.BasicFilter(expectedFilter.target, expectedFilter.operator, ...expectedFilter.values);
+      const filter1 = new models.BasicFilter(expectedFilter.target, expectedFilter.operator, expectedFilter.values);
+      const filter2 = new models.BasicFilter(expectedFilter.target, expectedFilter.operator, ...expectedFilter.values);
 
       // Assert
       expect(filter1.toJSON()).toEqual(filter2.toJSON());
@@ -463,14 +445,14 @@ describe("Unit | Filters", function () {
   describe("AdvancedFilter", function () {
     it("should throw an error if logical operator is not a non-empty string", function () {
       // Arrange
-      const condition: models.report.IAdvancedFilterCondition = {
+      const condition: models.IAdvancedFilterCondition = {
         value: "a",
         operator: "LessThan"
       };
 
       // Act
       const attemptToCreateFilter = () => {
-        return new models.report.AdvancedFilter({ table: "t", column: "c" }, <any>1, condition);
+        return new models.AdvancedFilter({ table: "t", column: "c" }, <any>1, condition);
       };
 
       // Assert
@@ -479,7 +461,7 @@ describe("Unit | Filters", function () {
 
     it("should throw an error if more than two conditions are provided", function () {
       // Arrange
-      const conditions: models.report.IAdvancedFilterCondition[] = [
+      const conditions: models.IAdvancedFilterCondition[] = [
         {
           value: "a",
           operator: "LessThan"
@@ -496,7 +478,7 @@ describe("Unit | Filters", function () {
 
       // Act
       const attemptToCreateFilter = () => {
-        return new models.report.AdvancedFilter({ table: "Table", column: "c" }, "And", ...conditions);
+        return new models.AdvancedFilter({ table: "Table", column: "c" }, "And", ...conditions);
       };
 
       // Assert
@@ -508,7 +490,7 @@ describe("Unit | Filters", function () {
 
       // Act
       const attemptToCreateFilter = () => {
-        return new models.report.AdvancedFilter({ table: "Table", column: "c" }, "Or", { value: "a", operator: "Contains" });
+        return new models.AdvancedFilter({ table: "Table", column: "c" }, "Or", { value: "a", operator: "Contains" });
       };
 
       // Assert
@@ -517,7 +499,7 @@ describe("Unit | Filters", function () {
 
     it("should output the correct json when toJSON is called", function () {
       // Arrange
-      const expectedFilter: models.report.IAdvancedFilter = {
+      const expectedFilter: models.IAdvancedFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -537,7 +519,7 @@ describe("Unit | Filters", function () {
       };
 
       // Act
-      const filter = new models.report.AdvancedFilter(
+      const filter = new models.AdvancedFilter(
         expectedFilter.target,
         expectedFilter.logicalOperator,
         ...expectedFilter.conditions);
@@ -548,7 +530,7 @@ describe("Unit | Filters", function () {
 
     it("can be constructed using either array form or individual arguments", function () {
       // Arrange
-      const expectedFilter: models.report.IAdvancedFilter = {
+      const expectedFilter: models.IAdvancedFilter = {
         $schema: "http://powerbi.com/product/schema#advanced",
         target: {
           table: "a",
@@ -568,8 +550,8 @@ describe("Unit | Filters", function () {
       };
 
       // Act
-      const filter1 = new models.report.AdvancedFilter(expectedFilter.target, expectedFilter.logicalOperator, expectedFilter.conditions);
-      const filter2 = new models.report.AdvancedFilter(expectedFilter.target, expectedFilter.logicalOperator, ...expectedFilter.conditions);
+      const filter1 = new models.AdvancedFilter(expectedFilter.target, expectedFilter.logicalOperator, expectedFilter.conditions);
+      const filter2 = new models.AdvancedFilter(expectedFilter.target, expectedFilter.logicalOperator, ...expectedFilter.conditions);
 
       // Assert
       expect(filter1.toJSON()).toEqual(filter2.toJSON());
@@ -580,20 +562,20 @@ describe("Unit | Filters", function () {
     it('getFilterType should return type of filter given a filter object', function () {
       // Arrange
       const testData = {
-        basicFilter: new models.report.BasicFilter({ table: "a", column: "b" }, "In", ["x", "y"]),
-        advancedFilter: new models.report.AdvancedFilter({ table: "a", column: "b" }, "And",
+        basicFilter: new models.BasicFilter({ table: "a", column: "b" }, "In", ["x", "y"]),
+        advancedFilter: new models.AdvancedFilter({ table: "a", column: "b" }, "And",
           { operator: "Contains", value: "x" },
           { operator: "Contains", value: "x" }
         ),
-        nonFilter: <models.report.IFilter>{}
+        nonFilter: <models.IFilter>{}
       };
 
       // Act
 
       // Assert
-      expect(models.report.getFilterType(testData.basicFilter.toJSON())).toBe(models.report.FilterType.Basic);
-      expect(models.report.getFilterType(testData.advancedFilter.toJSON())).toBe(models.report.FilterType.Advanced);
-      expect(models.report.getFilterType(testData.nonFilter)).toBe(models.report.FilterType.Unknown);
+      expect(models.getFilterType(testData.basicFilter.toJSON())).toBe(models.FilterType.Basic);
+      expect(models.getFilterType(testData.advancedFilter.toJSON())).toBe(models.FilterType.Advanced);
+      expect(models.getFilterType(testData.nonFilter)).toBe(models.FilterType.Unknown);
     });
   });
 });
